+++ src/cbc_sdk/__init__.py
@@ -1,10 +1,21 @@
+# *******************************************************
+# Copyright (c) Broadcom, Inc. 2020-2024. All Rights Reserved. Carbon Black.
+# SPDX-License-Identifier: MIT
+# *******************************************************
+# *
+# * DISCLAIMER. THIS PROGRAM IS PROVIDED TO YOU "AS IS" WITHOUT
+# * WARRANTIES OR CONDITIONS OF ANY KIND, WHETHER ORAL OR WRITTEN,
+# * EXPRESS OR IMPLIED. THE AUTHOR SPECIFICALLY DISCLAIMS ANY IMPLIED
+# * WARRANTIES OR CONDITIONS OF MERCHANTABILITY, SATISFACTORY QUALITY,
+# * NON-INFRINGEMENT AND FITNESS FOR A PARTICULAR PURPOSE.
+
 from __future__ import absolute_import
 
 __title__ = 'cbc_sdk'
 __author__ = 'Carbon Black Developer Network'
 __license__ = 'MIT'
 __copyright__ = 'Copyright 2020-2024 VMware Carbon Black'
-__version__ = '1.5.2'
+__version__ = '1.5.3'
 
 from .rest_api import CBCloudAPI
 from .cache import lru