+++ src/tests/unit/workload/test_compliance.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 # *******************************************************
-# Copyright (c) VMware, Inc. 2021-2024. All Rights Reserved.
+# Copyright (c) Broadcom, Inc. 2020-2024. All Rights Reserved. Carbon Black.
 # SPDX-License-Identifier: MIT
 # *******************************************************
 # *